--- src/components/atoms/index.js
@@ -1,1 +1,0 @@
-// export all here
